requirements.txt
streamlit
numpy
pandas
yfinance
scikit-learn
matplotlib
tensorflow
ta-lib-bin
scipy